Add handling of EINTR for ZmqPoll
If the process that runs ZmqPoll got a signal from a child process,
ZmqPoll is interrupted, leading to a break loop. We now handle this
case by just returning an empty list.

This is of course not optimal since we can't distinguish between a
timeout and EINTR, and this should be fixed in a future version.